BETA: Update Bitcode framework without x86 architectures
With the prior Bitcode beta, the Bitcode segment was present and Xcode didn't complian about Bitcode being missing from the TextExpander.framework, even so, Xcode merrily stripped away the Bitcode. This appears to have something to do with the x86 architectures being present, so we've removed them, at least for now.